Move blinded message path util into message submodule
